#97 rename SRSignalRConfig to SRLog
Couldn't get SignalR to auto detect CocoaLumberjack but this gets us
close.
If cocoa lumberjack is installed define COCOA_LUMBER_JACK in SRLog to
use it if not we have our own NSLog macros that will allow switching on
and off of logging levels.  From this perspective logging with or
without Cocoalumberjack appears no different.  The advantage of cocoa
lumberjack is it allows for logging to multiple locations.