Fixed server send event stream reader handler
If the stream is already opened by the time it's passed to the stream reader and getting NSStreamEventOpenCompleted event is in a race condition.

When NSStreamEventHasSpaceAvailable event is handled, it is safe to assume that the stream has already been opened and the reader state m
ust be updated accordingly. Otherwise the stream can never be read.