prevent longPolling from running forever
- when testing if you either timeout or abort connections (using
fiddler or some other mitm tool) on most connections, they fallback
regularly, but if you ever get to longPolling, it would continue
retrying forever - even if there no connection ever comes up. This
change checks if we are active when it gets an error in longPolling. If
outside the reconnectWindow, it will stop the connection.